requirements.txt
opencv-python
numpy
trimesh
Pillow
matplotlib
scikit-image
ifcopenshell
torch
torchvision
diffusers
transformers
accelerate
easyocr
streamlit
